requirements.txt
joblib
pyyaml
pandas
flask
tensorflow
numpy
scipy
logging